[objc] Add empty annotations that correspond to the captured variables added as formals of blocks
Summary:
We add captured as formals in ObjC blocks, but we were missing adding annotations (at least empty ones) for those new formals. This means that the annotations that we do have for the formals of blocks were not treated correctly, because the list of formals and annotations was mismatched.
This diff fixes it.

It also removes those dummy annotations for the captured in the Parameter Not Null Checked checker.

Reviewed By: skcho

Differential Revision: D32724769

fbshipit-source-id: f7a6ddab10