[report] sort issues from issue log together with the others
Summary:
Before we were first sorting the issues not in issue logs, then adding
those in issue logs, which meant that report.json wasn't sorted all the
way.

Now the sorting is done on all the issues at once.

This doesn't change the order of the issues in tests yet because
--issues-tests does its own sorting after loading report.json, which was
hiding this problem. See next diff.

Reviewed By: ngorogiannis

Differential Revision: D32677589

fbshipit-source-id: 3125c67287