[pulse] add support for floats
Summary:
Treat them as rationals, which is pretty straightforward as integers are
already treated as rationals. The main new thing is a way to pass float
constants to `PulseFormula`.

Treatment of float division vs int division in the next diff, and
treatment of float comparison vs int comparison TODO (pulse translates x
< y as x + 1 <= y, assuming ints; need more SIL support to distinguish
between semantics, or need to strengthen arithmetic domain so that "x <=
y /\ x ≠ y" is treated as well as the current semantics).

Closes https://github.com/facebook/infer/issues/1555

Reviewed By: ezgicicek

Differential Revision: D32465565

fbshipit-source-id: 5ec8d8374e