[paramNotNull] Simplify method name in traces
Summary:
Simplifying the method name in traces, in particular for blocks given that we can tell which block because of the line number, a simple `block` is enough, the current complex name is not very readable.

This also removed the class name from the method name in ObjC methods, which also looks more readable, given the line number shows which method it is.

Reviewed By: ngorogiannis

Differential Revision: D33404783

fbshipit-source-id: ab097bf6f6