[pulse][retain-cycle] detect and report retained cycles
Summary:
There is a retain cycle if there is a strong reference path from a
variable to itself. The detection chooses the safe solution of assuming
a reference not to be strong when it is not able to determine its
nature to avoid getting too many FP

The new `RefCounted` attribute works in parallel with the `Allocated`
one similarly with how the `RetainCyle` issue does with `MemoryLeak`. It
indicates a value that is handled via a GC relying on reference counting
to determine when a variable should be freed/deallocated/destroyed...
In the current context, it indicates an objective-c variable handled via
ARC but it could also be used for e.g. `shared_ptr` in C++

Reviewed By: dulmarod

Differential Revision: D32729817

fbshipit-source-id: ec449ac09c