[pulse][retain-cycle] avoid removing retained values
Summary:
In objective-c, memory is managed via ARC (automatic reference counting)
which relies on compiler-injected `retain` and `release` calls to
increase or decrease a ref count and decide when a variable's `dealloc`
method can/should be called. We emulate this reference counting system
in pulse by counting all references before any call to `dealloc` and
`dealloc` does the work of knowing of it should actually do the whole
deallocation now or in a later call (emulalting `release` in way).

Those `dealloc` calls are injected whenever a variable is seen in a
`Metadata ExitScope` instruction. However, some may be locked in a
retain cycle at this point and released from it later but it will only
be seen in an `ExitScope` once because variables in an `ExitScope` are
removed from the stack (they will not be read or written anymore). This
leads to variables never being properly deallocated and therefore, can
lead to FP memory leaks.

To fix it, we avoid removing variables that will not be deallocated
during their `ExitScope` and defer the deallocation and stack removal.
At the end of a function, we consider all the variables in the stack
which are not return variables as exiting the scope, offering them a
second and last chance to be deallocated

Reviewed By: dulmarod

Differential Revision: D32836480

fbshipit-source-id: f6b6e5ed30