Fixing liveness wrt to exceptions
Summary:
The liveness analysis is the only anaysis that relies on BackwardCfg but it is unfortunatly unsound wrt Java exceptions. We provide a new backward abstract interpreter and instatiate it ofr liveness.

On the following example
```
class Liveness {

    public A main() {
        A a = null;
        try {
            a = new A();
            return null;
        } catch (IOException e) {
            return a;
        }
    }

}

class A {

    public A() throws IOException {
        throw new IOException();
    }

}
```
the instruction `a = new A();` generates a block
```
n$0 = new(sizeof(A));
n$1 = A.<init>(n$0);
a = n$0
```
but the liveness predict that `a` is dead at the entrance of the block. This is wrong if `A.<init>`may throw an exception and the corresponding handler uses `a` before redefining it.

Indeed, on a blocks with outgoing exceptional edges, the backward CFG is not really symetric to the forward.

```
Forward block:
instr1; --> outgoing exn edge
instr2; --> outgoing exn edge
intrs3; --> outgoing exn edge
--> outgoing normal edge
```
```
Backward block:
instr1; <-- ingoing exn edge
instr2; <-- ingoing exn edge
intrs3; <-- ingoing exn edge
<-- ingoing normal edge
```
Hence the ingoing live sets must be merged at **each** instruction in the block. The merge is performed before the instruction because the potential assignement is not done if an exception is thrown.

Reviewed By: skcho

Differential Revision: D33548844

fbshipit-source-id: 504b34e75f