[java] enforce struct invariants
Summary:
The Java frontend sometimes produces structs with duplicate fields, procedures or super types.  This diff enforces uniqueness and sorts all lists, so that this invariant can be exploited when type environment merging.

NB only Java types are affected.  Even though the clang frontend has many of the same issues, changing the order of fields is problematic because of anonymous initializers. Also, there is no type environment merging for clang.

Reviewed By: ezgicicek

Differential Revision: D33623223

fbshipit-source-id: cc5fbacd2f