[pulse][erlang] Configurable custom Erlang models.
Summary:
This adds the option `--pulse-models-for-erlang`, which is meant to be
provided via a JSON object in .inferconfig. The value of this option
should be a list of rules, where each rule is a (JSON) object with two
fields, selector and behavior. Currently, there is only one selector
("MFA", which stands for "model, function, arity") and one behavior
("ReturnValue", which can be used to build atoms, integers, lists,
tuples).

Differential Revision: D32994361

fbshipit-source-id: 43a3512b61