[topl] Apply limits after exec steps, not before
Summary:
Topl has two main functions that evolve its abstract state: `small_step`
and `large_step`. Both used to start by applying limits (on the size of
the abstract state) and then continue by computing the transfer
function. As a result, the returned abstarct state may have been over
the limit. This was a problem because some of these returned states
ended up being summaries of procedures. And, coupled with the fact that
summaries were never trimmed down to limits, this led to an explosion
that led to OOMs.

The fix in this diff is simple: apply the limits *after* applying the
transfer function. (A fancier fix would be to apply a hard limit, as you
go. But that doesn't seem necessary.)

Reviewed By: mmarescotti

Differential Revision: D34454958

fbshipit-source-id: 89041e20f6