[pulse] uninit cannot be non-fatal and latent
Summary:
Issues cannot be both non-fatal and latent because that would require
both continuing the execution and stopping it to create a
`LatentAbortProgram` summary. What was happening when applying these
summaries was that we would apply the latent issue pre/post, then turn
it into a summary (thereby forgetting all the local variables, etc.!),
then continue the execution from that summary. Summaries should never go
through `exec_instr` (because they are missing all the local state) so
this was a bug.

Reviewed By: skcho

Differential Revision: D34548773

fbshipit-source-id: 9e4bb56836