[pulse][block] only specialize when needed
Summary:
The previous analysis-time specialization was very agressive and tried
to specialize every function called. Instead, we want to try to
specialize only functions with some missing information during the
analysis. This missing information is detected when we try to call an
unkown block. The function will indicate to its caller that it needs to
be specialized and if it still doesn't succeed in calling the block
after that, then the caller will tell its caller that it needs to be
specialized and so on until we succeed or nobody can be specialized
anymore.

Reviewed By: jvillard

Differential Revision: D34254078

fbshipit-source-id: b8a1f696cb