[pulse] only suppress null deref errors when there are skipped calls
Summary:
Currently, we suppress all Java issues where unknown functions were
called on the way, except for some list of functions such as `<init>`.
This was done back when only detected null errors on Java but that's no
longer the case, so restrict this filtering to just null issues.

As an aside, this is probably too strict even for null issues, we may
want to get rid of this filter altogether in the future.

Reviewed By: da319

Differential Revision: D34587654

fbshipit-source-id: 339d043beb