[pulse] Add models for functions returning a copy
Summary:
This diff adds models for functions that return a copy.

Context: We don't currently track unnecessary copies inter-procedurally. Hence, for functions that are known to return copies (`folly::get_default` or `folly::get_optional`), we have no way of detecting the copy because there is no explicit call to a copy ctor. In this diff, we introduce models for detecting such functions.

Reviewed By: skcho

Differential Revision: D33978971

fbshipit-source-id: ad0be68d6d