[infer] Mark modifications before destructors are called
Summary:
We were checking for modifications to the copy at ExitScope only but this resulted in false-positives due to destructors getting rid of the addresses associated with the copied var as in the following test:

```
 auto copy = source_vec; // copy tracked successfully
 copy[0] = 0; // copy modified
 //1) call to destructor: at this point, addresses copy points to are removed from the heap completely
 //2) ExitScope: we compared source heap with an empty heap and concluded no modification
```

To deal with these cases, we should check for modifications before the call to the destructor rather than at ExitScope.

If something is checked for modifications before the destructor call, we shouldn't also check for it at ExitScope since it is possible that the destructor modified the copy. To deal with this, we extend the non-disj. domain to a pair where the second component keeps track of all the destructor-checked variables. I wonder if there is a better way to do that?

Reviewed By: skcho

Differential Revision: D34683950

fbshipit-source-id: aa30828e6e528cdf49b9907c9955b63f767c235e