[java][frontend]Fixing try_with_resource new compiling schema
Summary:
The version 11 of `javac` does not produce the same bytecode programs than version 8. In particular the try-with-resource construction is not compiled the same way. The main difference lies in the exception table that is generated.

Version 8
```
 Exception table:
         from    to  target type
            63    67    70   Class java/lang/Throwable
            19    55    86   Class java/lang/Throwable
            19    55    91   any
           100   104   107   Class java/lang/Throwable
            86    92    91   any
```

Version 11
```
 Exception table:
         from    to  target type
            17    40    47   Class java/lang/Throwable
            48    52    55   Class java/lang/Throwable
```

The `any` type corresponds to `handler.JBir.e_catch_type = None` in Sawja.
Since `java.lang.Thowable` is the highest class in the hierarchy of Exception, we can reuse the same code than for `any`.

Reviewed By: jvillard

Differential Revision: D34790074

fbshipit-source-id: 848be1ccc54097d3a1f2ff2567fa66c838d441a3