[block] only convert blocks and specialized function calls into closures
Summary:
We were converting calls like `[obj initWithCallback:callback]` when
`callback` is a known block into closure calls although there is no
specialization for `init` functions. This lead to closure calls with
ignored captured variables in Pulse because the nature of the function
to be called is verified before linking the expected and given captured
vairables. However, other analyzers may not do this verification with
reasons.
Ignoring the captured variables is not the expected behavior in this
context. Rather, we should not have any in the first place.

With the added condition, we get rid of the undesired closure
conversions

Reviewed By: dulmarod

Differential Revision: D34862066

fbshipit-source-id: 5e136decc496422d0a75ebbf6d841049cbb6026a