Have make test print a replace command from the root dir
Summary:
When running the tests, when the output of infer changes, make will
print a command that replaces the expected output with the actual
one. This command includes a relative path which we compute from the
root dir rather than the test one, because tests are usually run from
this root dir.

Reviewed By: jvillard

Differential Revision: D35360628

fbshipit-source-id: cfc47767f33e51d42701f1323633bc434366c00b