[pulse] import quandary/ExternalSpecs.java test, delay interprocedural taint checks
Summary:
This revealed an issue with checking taint during function calls: we
were checking for taint after applying the precondition but at this
point sanitizers applied during the call are not visible yet as they are
the post's attributes.

Also made adjustments to .inferconfig to make the tests pass:
- "procedure" does not attempt to walk the class hierarchy, not sure we
  want to bother so switch to matching using "class_names" / "method_names"
  instead when we care about that.
- the index of the taint in a constructor is now 1 instead of 0, I guess
  quandary had some logic to avoid exposing infer's encoding of the
  object layer (with `this` at position 0) to the user but we don't yet
  (not sure if we should either).

There were two FNs in this file for quandary which, pleasingly, pulse
does not suffer from.

Reviewed By: skcho

Differential Revision: D35360871

fbshipit-source-id: 540fa68bc0cad65461d53aa917d3f174271ee564